Implement package resolution.
This commit implements a file system package resolver that will load foreign
packages from either a binary-encoded wasm file or a directory containing WIT
files.

The package is parsed and the wasmparser types converted to the WAC
representation, allowing the resolution of package references from already
implemented AST resolution:

* import statements
* use statements in an interface or world.
* world include statements.
* world item statements.

The `as` clause in use statements is also implemented as part of these changes.

The type system has also been refactored to remove the lifetime annotations.