Remove `smallvec` dependency.
Just use an array of `Token` now that tokens are only a single byte (they used
to be much larger before a previous refactoring).

Additionally, add test coverage for multiple expected tokens errors.

Fix a bug where a failed lookahead used the current token position instead of
the last position in the source where the next token is EOI.