Implement export statement resolution.
This commit completes AST resolution.

An `imports` and `exports` collection were added to the resolution to keep
track of imported and exported items from the composition.

The resolution now ensures that import names are unique.

Additionally, this updates to the latest dependencies including changes to
`wasmparser`'s type information representation.

Import and export names are now validated in accordance with the latest
component model spec; this will enable specifying import dependencies when
performing encodings.