Implement support for registry packages.
This commit implements support for referencing packages from a registry.

A new `wac-resolver` crate was added and the resolvers implementation was
refactored into it; package resolution now occurs before document resolution.

A `registry` feature was added to the CLI tool; when enabled, a `--registry`
CLI option is accepted to point at the registry to use.

The CLI also supports Warg client configuration files (default or local).

A simple `indicatif` progress bar was added to display the number of packages
that need to be downloaded from the registry.

A new integration test that uses packages from a registry was also implemented.