Fix merge of disjoint interfaces.
Currently, a bug exists in merging of interfaces such that if one interface
exports a function that references an exported (i.e. "named") type and another
interface exports a different function that references the same exported type,
the merged interface would have references to function types with types that
are not present in the "encoded type index" map.

This results in the encoder re-encoding the same type and using that newly
encoded type index instead of the correct type index to the named type.

Ultimately this resulted in a validation error of the encoded output where the
interface wasn't valid for import because of the function not using what should
be a named value type.

The fix is for the merged interface to maintain a mapping between the new type
and the old types so that the "encoded type index" map is updated accordingly.