Fix duplicate import encoding.
This fixes a bug where a duplicate import might occur when an import with
dependencies on other interfaces is encoded and then later on an import for one
of those dependencies is also explicitly encoded.

It skips the import if it has already been imported and also ensures that the
full interface definition is provided when encoding at the root state (i.e. for
the composition itself and not an instance/component type).

As such, we don't need to process dependencies of implicit imports during
resolution.