Update README and fix incorrect error message.
This commit updates the README in anticipating of moving the project into a BA
repository.

It creates a separate `LANGUAGE.md` to document the WAC language and includes
usage examples of the `wac` CLI in `README.md`.

Also included in this change is a fix to the error message that occurs when an
export name cannot be inferred (it should have been updated with the `with`
keyword removal).