Implement spreading of instantiation args and exports.
This commit implements a syntax for "spreading" instantiation arguments and
exports.

For instantiation, a `...<id>` syntax is now supported in an instantiation
argument list where `<id>` names an instance; this will automatically alias and
pass any matching exports of the instance as arguments to the instantiation.

Similarly, for exports, a `<expr>...` syntax is now supported for automatically
exporting the exports of an instance from the composition.

Also fixed:

* Incorrect parsing of comma-delimited tokens (comma was unintentionally
  optional).
* Confusing spans on some errors relating to expressions; expressions now keep
  track of their spans for better error messages.