Improve subtype check.
This commit improves the subtype checking logic and the error messages it
produces.

It fixes a bug where the `targets` check was not properly performing checks
against imports, which should be contravariant.

Also fixes a warning relating to a deprecated function call; the fix is to not
remove from the collection as we no longer have to since the bytes of a package
are behind an `Arc`.