Implement a composition graph library.
This commit implements a composition graph library that will serve as the basis
for how WAC does AST resolution in the near future.

The inspiration comes from the existing `CompositionGraph` API in
`wasm-compose`, except this API is more flexible, allowing a composition graph
to define its own types, have explicit exports, and export any node in the
graph.

The type representation present in `wac-parser` has been extracted to its own
crate, `wac-types`. It will be shared in the future between the parser and the
graph library.

As a result, a chunk of this new code is just copied out of `wac-parser`. Once
the graph library is merged, a follow-up commit will eliminate the duplicated
code and change `wac-parser` to resolve ASTs to a `CompositionGraph`.