Reimplement `wac-parser` on `wac-graph`.
This commit is a complete reimplementation of `wac-parser`'s resolution to be
based on `wac-graph`.

It also contains an overhaul of the `wac-graph` API to facilitate its use from
`wac-parser`:

* Reimplements `CompositionGraph` on top of `petgraph::StableDiGraph`, which
  eliminates needing to keep track of nodes internally to `CompositionGraph`.
* Packages now populate their type information into the graph's types
  collection rather than having their own collections.
* Splits the `Error` enum into discrete error types for each graph operation;
  this allows the AST resolver to properly attach span information to errors.
* Add type dependency edges for defined types, ensuring topological ordering.
* Fix encoding of defined type aliases so that each encoded alias refers to the
  export index of the previously encoded type.
* Added a debug formatter for `CompositionGraph` that outputs a DOT
  representation (used in resolution tests as well).
* Graph encoding now includes an optional producers custom section.
* Fixed toposort to be in ascending node index order for independent nodes.

The subtype checker implementation was reverted to the previous implementation,
where `invert` and `revert` are called at specific points to change the type of
check being performed; the newer implementation did not give the correct
"expected, found" error messages for AST resolution.

`wac resolve` now outputs a DOT representation of the resolved composition
graph instead of JSON.

Most of the deleted code comes from code that has already moved into
`wac-types` and `wac-graph`.

Closes #79.
Fixes #78.
Fixes #76.