Keep better track of already seen instances in encoder.
Before this change, the encoder would not keep track of instances that
were not in the implicit instantiation args but were still deps of
some other instances and thus imported.

With this change, instance tracking is moved inside of the encoder so it
always knows when it encodes an instance to refer to that encoded
instance if the instance is encountered again instead of re-encoding.