Fix duplicate imports that may occur from implicitly imported items.
This commit refactors the implicit imports so that it uses the existing
`import` method, which already checked for duplicates and did the proper
bookkeeping in the encoding state.

Additionally, state now keeps track of encoded instances based on the
interface names (i.e. `id`) rather than the arena id; this corresponds to a
previous change in the aggregator for merging interface definitions together.

As there's no longer a single types arena, logging the ids of things during
encoding just led to confusing; this commit also audits the log messages from
`wac_graph` and fixes them where appropriate.