Fix the encoding of imported resources in the `CompositionGraph`.
This commit fixes the encoding of imported resources in the `CompositionGraph`.

Previously a panic was encountered when the graph attempts to import, either
explicitly or implicitly, a resource.

The fix requires tracking the "owner" of an aliased resource so that the export
on the owning interface may be aliased and used for type equality when encoding
the resource import.

Fixes #62.