Fix infinite recursion caused by aliased resources.
This commit fixes an existing bug exposed by the refactoring (likely due to the
changes in the order in which nodes are processed for implicit arguments).

Previously, an interface might contain a type alias for another resource it
owns; this lead to the alias information of the resource being self-referential
for the interface containing the aliased resource.

It would then lead to infinite recursion in an attempt to remap such an
interface when populating implicit instantiation arguments.

The fix is to clear the self-referencing information while converting an
interface during package parsing.