Add a test for a dependency issue
The issue is that dependencies of implicit imports and explicit imports
are not merged and thus if you've already imported a dependency of an
implicit import, the dependency of the explicit import will effectively
be ignored. This is an issue when the explicit import dependency has
different items than the version imported as part of the implicit
import.